Create utterances.json
Hello Maria, We are starting to look to onboard Activity Tracker into Chatbot (as we did already with LogAnalysis)  If you could please approve and merge this utterance file we would appreciate it, following it being pushed to production we will then build the intents in Chatbot. Please reach out with any questions, Claire Thank you!